requirements.txt
Flask
flask-restful
flask-cors
gevent
gunicorn
python-dotenv
psutil
Flask-Swagger-UI
scikit-learn
pyzmq